+++ include/ArduinoJson/Polyfills/isFloat.hpp
@@ -8,12 +8,13 @@
 #pragma once
 
 #include "./ctype.hpp"
-#include "./math.hpp"
 
 namespace ArduinoJson {
 namespace Polyfills {
 
 inline bool isFloat(const char* s) {
+  if (!s) return false;
+
   if (!strcmp(s, "NaN")) return true;
   if (issign(*s)) s++;
   if (!strcmp(s, "Infinity")) return true;